Beam position monitor (BPM) device implementation improved
The variables of the beam position monitors can now
be accessed using the single variable 'packed data'.
The values in the 'packed data' are also provided as
access for each bpm. These can be accessed using
the module bpm_single.

The bpm readings can be plotted for both using
examples/bpm_comparison.py

The following changes were made to the different
modules:

* ophyd/devices/pp/bpm/bpm.py
  added comment and deleted trailing withspace
* ophyd/devices/utils/derived_signal.py
  Separate class for conversion of BPM signals to
  physical coordinates
* ophyd/devices/pp/bpm/bpm_parameters.py
  reference to the conversion of bpm readings to
  physical coordinates
* ophyd/devices/pp/bpm/bpm_single.py
  refactored channel code using FormattedComponent

* examples/bpm_reading.py
  different set of lists of plots.

* bluesky/live_plot/bpm_plot
  Plotting bpm readings versus ds and using the bpm
  names as labels.

* examples/bpm_comparison.py
  Plots bpm data as provided by bpm and bpm_single